Resolving the review comments raised for notification-module pull-request.
